+++ CHANGELOG.md
@@ -6,4 +6,4 @@ All notable changes to this project will be documented in this file.
 - UI to show and set channel topics.
 - Display currently connected peer count.
 - Settings screen for future features and a button to remove the cabal from the client.
-- Slash commands for `/help`, `/nick`, `/topic`, `/join` (for switching channels), and `/remove` (for removing a cabal from the client).
+- Slash commands for `/help`, `/join`/`/j`, `/motd` (message of the day), `/nick`/`/n`, `/remove`, `/topic`, and `/remove` (for removing a cabal from the client).